Added Recording Threshold, New comments
Added a recording threshold for sampling percussive sounds. You can now hold the record button and the record function will not start until the input signal reaches a level slightly above the noise floor. The noise floor for a 12bit A/D is typically around -66db, so the threshold is around -50db or so.

Added commentary describing the hardwired passthrough mute button.